+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.0.1
 
 - Support additional bundler arguments to be passed when using docker (#2)
+- Support multi-value query string parameters
+- Support multi-value headers in request and response
 
 # 1.0.0